+++ Examina/Services/IStudentExamService.cs
@@ -151,7 +151,7 @@ public interface IStudentComprehensiveTrainingService
     /// <param name="durationSeconds">用时（秒，可选）</param>
     /// <param name="notes">备注（可选）</param>
     /// <returns>是否标记成功</returns>
-    Task<bool> MarkTrainingAsCompletedAsync(int trainingId, decimal? score = null, decimal? maxScore = null, int? durationSeconds = null, string? notes = null);
+    Task<bool> MarkTrainingAsCompletedAsync(int trainingId, double? score = null, double? maxScore = null, int? durationSeconds = null, string? notes = null);
 
     /// <summary>
     /// 获取综合训练完成记录
