+++ ExaminaWebApplication/Models/Api/Student/StudentExamDto.cs
@@ -70,7 +70,7 @@ public class StudentExamDto
     /// <summary>
     /// 及格分数
     /// </summary>
-    public int PassingScore { get; set; }
+    public double PassingScore { get; set; }
 
     /// <summary>
     /// 是否随机题目顺序
@@ -156,7 +156,7 @@ public class StudentSubjectDto
     /// <summary>
     /// 权重
     /// </summary>
-    public decimal Weight { get; set; }
+    public double Weight { get; set; }
 
     /// <summary>
     /// 题目数量
